Move ScalingIndication to separate file
